refactor uid to long and rename it

+++ app/src/main/java/jasper/wagner/smartstockmarketing/domain/model/Stock.kt
@@ -14,5 +14,5 @@ data class Stock(
     @ColumnInfo(name = "lastTimeStamp") var lastTimeStamp: String
 ): Serializable {
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = "stockUID") var stockUID: Int = 1
+    @ColumnInfo(name = "stockUID") var stockUID: Long = 1
 }
