+++ app/src/main/java/jasper/wagner/smartstockmarketing/util/NotifyWorker.kt
@@ -26,7 +26,7 @@ class NotifyWorker(@NonNull context: Context, @NonNull params: WorkerParameters)
     @NonNull
     override fun doWork(): Result {
         val paramsString = inputData.getString(API_CALL_PARAMS)
-        val stockUID = inputData.getInt(STOCK_UID,0)
+        val stockUID = inputData.getLong(STOCK_UID,0)
         val growthMargin = inputData.getDouble(GROWTH_MARGIN,1.0)
         val channelID = inputData.getInt(NOTIFICATION_ID,100)
         val apiParams = SerializeHelper.deserializeFromJson(paramsString!!) as StockApiCallParams
